Moved Validator to shared module
