+++ collect_app/src/main/java/org/odk/collect/android/preferences/screens/ServerPreferencesFragment.java
@@ -53,7 +53,7 @@ import org.odk.collect.android.permissions.PermissionsProvider;
 import org.odk.collect.android.utilities.PlayServicesChecker;
 import org.odk.collect.android.utilities.SoftKeyboardController;
 import org.odk.collect.android.utilities.ToastUtils;
-import org.odk.collect.android.utilities.Validator;
+import org.odk.collect.shared.Validator;
 import org.odk.collect.shared.Md5;
 
 import java.io.ByteArrayInputStream;
