+++ shared/src/main/java/org/odk/collect/shared/Validator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.odk.collect.android.utilities;
+package org.odk.collect.shared;
 
 import java.util.regex.Pattern;
